pyproject.toml
[project]
name = "mrmd"
version = "0.1.0"
description = "Markdown that runs - AI-native literate programming"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Maxime Rivest" }
]
keywords = ["markdown", "repl", "literate-programming", "ai", "notebook"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "brepl>=0.1.0",
    "aiohttp>=3.9.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "ipython>=8.37.0",
    "black>=25.11.0",
    "watchdog>=3.0.0",
    "pycrdt>=0.10.0",  # Yjs-compatible CRDT for real-time collaboration
]

[project.optional-dependencies]
ai = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mrmd = "mrmd.cli.main:main"

[project.urls]
Homepage = "https://mrmd.dev"
Repository = "https://github.com/maximerivestio/mrmd"
Documentation = "https://mrmd.dev/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mrmd"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.uv.sources]
brepl = { path = "../brepl", editable = true }

[tool.uv.workspace]
members = [
    "ai-server",
]
